% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents location.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{single integer state code in the USA. Function will throw an
error in case no data is available for selected state and year.
In case you input a vector only first element, if data exists, will
be plotted and function will throw a warning.}

\item{year}{single integer value of the year that will be plotted}
}
\value{
a map object of the selected us state with points in the accident
        location of selected year.
}
\description{
This function creates a map that has a point in each accident location.
fars data files must be in working directory.
}
\examples{
\dontrun{
# plot existing region
fars_map_state(1, 2013)

# function throws error as data is not available
fars_map_state(2, 2013)

# map is plotted but function throws warning
fars_map_state(c(1,3), 2013)
}
}
\references{
\url{http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
\seealso{
\code{\link{map}}
}

