% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_read}
\alias{fars_read}
\title{Load fars data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{single path to file containing fars data, just filename if
data is stored in the same directory.
If file does not exist function will throw an error.}
}
\value{
a tibble data frame object containing fars data
}
\description{
This function reads fars data from *.csv.bz2 file, data is saved as a
tibble data frame object.See \code{\link{tbl_df}} for details.
}
\examples{
\dontrun{
# save file in folder named data in working directory
fars_file <- file.path("data", "accident_2014.csv.bz2")
fars2012 <- fars_read(fars_file)
}

}

