% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extract month and year variables.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{an integer or numeric vector containing years.}
}
\value{
a list of tibbles that include data for each of the input years.
}
\description{
This function extract month and year data from each file and stores it in
a tibble. It throws a warning for any year on which no data file is
available.
}
\examples{
\dontrun{
# this one return a one element list
fars_records_2013 <- fars_read_years(2013)

# this one return a two elements list and one warning as 2012 is not available
fars_2012_2014 <- fars_read_years(2012:2014)
}

}

