% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Number of accidents by month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{an integer or numeric vector containing years.}
}
\value{
a tibble data frame object containing number of accidents by
month and year.
}
\description{
This function computes number of accidents by months for years
selected as input. Function will throw a warning for each of the input years
that don't have data available.
fars data files must be working directory.
}
\examples{
\dontrun{
# summarise one year of data
monthly_summ_2013 <- fars_summarize_years(2013)

# summarise three years of data
monthly_summ_2013_2015 <- fars_summarize_years(2013:2015)
}

}
\references{
\url{http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}

