% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_filename}
\alias{make_filename}
\title{Build file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer or numeric vector containing years.}
}
\value{
character vector of the same size containing fars data files names.
}
\description{
This function builds file name string of input years. This function does
not check whether files actually exist or not. This function assumes files
are located in working directory for other functions to work.
}
\examples{
\dontrun{
fars_File_2012 <- make_filename(2012)
fars_files_2012_2016 <- make_filename(2012:2016)
fars_files_not_conseq <- make_filename(c(2012,2014))
}

}

